pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "async-swe-agents"
version = "0.1.0"
description = "Asynchronous multi-agent delegation for software engineering tasks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "openhands-sdk==1.11.0",
    "openhands-workspace==1.11.0",
    "openhands-tools==1.11.0",
    "openhands-agent-server==1.11.0",
    "litellm==1.81.11",
    "pydantic==2.12.5",
    "fire==0.7.1",
    "rich==14.2.0",
    "pyyaml==6.0.3",
    "numpy==2.4.1",
    "datasets==3.0.1",
    "openai==2.8.1",
    "anthropic==0.74.1",
    "jiter==0.12.0",
    "tokenizers==0.21.4",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.10",
    "plotly>=6.0",
    "httpx>=0.28",
]
dev = [
    "pytest>=9.0",
    "ruff>=0.14",
]

[tool.setuptools.packages.find]
include = ["core*", "tasks*", "judge*", "config*"]
